requirements.txt
discord.py==1.7.3
dynamic-yaml==1.2.3
expiringdict==1.2.1
fuzzywuzzy==0.18.0
mongoengine==0.23.1
requests==2.25.1
simpleeval==0.9.10
